Refactor 'host', 'cve', 'service', and more entities into its own file
