+++ app/importers/NmapImporter.py
@@ -19,8 +19,12 @@ import sys
 
 from PyQt5 import QtCore
 
-from db.database import hostObj, note, osObj, serviceObj, portObj, l1ScriptObj
+from db.database import note, l1ScriptObj
+from db.entities.host import hostObj
 from db.entities.nmapSession import nmapSessionObj
+from db.entities.os import osObj
+from db.entities.port import portObj
+from db.entities.service import serviceObj
 from parsers.Parser import Parser
 from ui.ancillaryDialog import ProgressWidget, time
